requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
psutil
